// src/pages/smartReplenishment/pages/arrivalDetails/columns.d.ts
// 请确保此处与真实字段完全匹配
export enum keys {
  goodsName = 'goodsName',    // 商品名称
  goodsCode = 'goodsCode',    // 商品编码
  replenishmentCount = 'replenishmentCount',    // 补货数量
  replenishmentRatio = 'replenishmentRatio',    // 到货数量占比
  sapSatisfyCount = 'sapSatisfyCount',          // SAP满足数量
  outCount = 'outCount',                      // 装箱出库数
  receivingCount = 'receivingCount',          // 收货量
  acceptanceCount = 'acceptanceCount',        // 验收量
  differenceQuantity = 'differenceQuantity',   // 差异量

  billCode = 'billCode',                // 单据编号
  rowNum = 'rowNum',                    // 行号
}